\name{tyroid}
\docType{data}
\alias{tyr1}
\alias{tyr2}

\title{Tyroid cancer data}
\description{
  These are subsets of RMA normalized GEO datasets GSE3467 and
  GSE3678. Each containes the expression of 95 genes selected by t-test
  significance. The data is utilized in the tyroid demo.
}
\usage{data(tyr1)}
\format{List containing normalized data (`cdata' slot) and class
  labels (`cls' slot).}
\source{"http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE3467", "http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE3678"}

\keyword{datasets}
